Create utils method buildLink
Also updated EggspresLink component to use buildLink. We will use buildLink to process links in other components (e.g. in sidebar, author cards, etc.)
