+++ app/_components/EggspressLink.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import React from 'react'
 import getSlugs from './getSlugs'
-import { getEggspressSettings } from '../utils';
+import { buildLink } from '../utils';
 
 interface EggspressLinkProps {
   href: string;
@@ -55,43 +55,7 @@ const processInternalUrl = async (url: string): Promise<string> => {
 }
 
 const processExternalUrl = async (url: string): Promise<string> => {
-  const linkSettings = await getEggspressSettings('links')
-  const re = /:\/\/([^\/]*)/;
-  const match = url.match(re)
-
-  if ( (match && match[1]) ) {
-    const baseUrl = match[1]
-    for ( let i = 1; i <= 20; i++ ) {
-
-      if (!linkSettings[`modifyLinkBaseUrl${i}`]) {
-        continue
-      }
-
-      let isMatch = baseUrl.toLowerCase().includes(linkSettings[`modifyLinkBaseUrl${i}`].toLowerCase())
-
-      if (linkSettings[`modifyLinkStrictMatch${i}`]) {
-        isMatch = linkSettings[`modifyLinkBaseUrl${i}`] === baseUrl
-      }
-
-      if ( isMatch ) {
-        let newUrl = 'https://'
-        if (linkSettings[`modifyLinkSetPrefix${i}`]) {
-          newUrl += linkSettings[`modifyLinkSetPrefix${i}`]
-        }
-        if (linkSettings[`modifyLinkSetNewBaseUrl${i}`]) {
-          newUrl += linkSettings[`modifyLinkSetNewBaseUrl${i}`]
-        } else {
-          newUrl += baseUrl
-        }
-        if (linkSettings[`modifyLinkSetSuffix${i}`]) {
-          newUrl += linkSettings[`modifyLinkSetSuffix${i}`]
-        }
-        return newUrl
-      }
-    }
-  }
-
-  return url
+  return await buildLink(url)
 }
 
 const EggspressLink: React.FC<EggspressLinkProps> = async ({href, id, children}: EggspressLinkProps) => {
