+++ app/utils.tsx
@@ -175,7 +175,46 @@ export async function getColors(prefix: string, colorKey: string, fallbackDark='
   }
 
   return classNames.join(' ')
+}
 
+export async function buildLink(url: string) {
+  const linkSettings = await getEggspressSettings('links')
+  const re = /:\/\/([^\/]*)/;
+  const match = url.match(re)
+
+  if ( (match && match[1]) ) {
+    const baseUrl = match[1]
+    for ( let i = 1; i <= 20; i++ ) {
+
+      if (!linkSettings[`modifyLinkBaseUrl${i}`]) {
+        continue
+      }
+
+      let isMatch = baseUrl.toLowerCase().includes(linkSettings[`modifyLinkBaseUrl${i}`].toLowerCase())
+
+      if (linkSettings[`modifyLinkStrictMatch${i}`]) {
+        isMatch = linkSettings[`modifyLinkBaseUrl${i}`] === baseUrl
+      }
+
+      if ( isMatch ) {
+        let newUrl = 'https://'
+        if (linkSettings[`modifyLinkSetPrefix${i}`]) {
+          newUrl += linkSettings[`modifyLinkSetPrefix${i}`]
+        }
+        if (linkSettings[`modifyLinkSetNewBaseUrl${i}`]) {
+          newUrl += linkSettings[`modifyLinkSetNewBaseUrl${i}`]
+        } else {
+          newUrl += baseUrl
+        }
+        if (linkSettings[`modifyLinkSetSuffix${i}`]) {
+          newUrl += linkSettings[`modifyLinkSetSuffix${i}`]
+        }
+        return newUrl
+      }
+    }
+  }
+  
+  return url
 }
 
 // Function below requires Node 20, which AWS Lambda does not support-- see above function for glob implementation
